Move Scene.update_frame()
