+++ manim/renderer/cairo_renderer.py
@@ -1,4 +1,6 @@
 import numpy as np
+from .. import file_writer_config
+from ..utils.iterables import list_update
 
 
 class CairoRenderer:
@@ -6,6 +8,46 @@ class CairoRenderer:
         self.scene = scene
         self.camera = camera
         self.file_writer = file_writer
+
+    def update_frame(  # TODO Description in Docstring
+        self,
+        mobjects=None,
+        background=None,
+        include_submobjects=True,
+        ignore_skipping=True,
+        **kwargs,
+    ):
+        """Update the frame.
+
+        Parameters
+        ----------
+        mobjects: list, optional
+            list of mobjects
+
+        background: np.ndarray, optional
+            Pixel Array for Background.
+
+        include_submobjects: bool, optional
+
+        ignore_skipping : bool, optional
+
+        **kwargs
+
+        """
+        if file_writer_config["skip_animations"] and not ignore_skipping:
+            return
+        if mobjects is None:
+            mobjects = list_update(
+                self.scene.mobjects,
+                self.scene.foreground_mobjects,
+            )
+        if background is not None:
+            self.camera.set_frame_to_background(background)
+        else:
+            self.camera.reset()
+
+        kwargs["include_submobjects"] = include_submobjects
+        self.camera.capture_mobjects(mobjects, **kwargs)
 
     def get_frame(self):
         """
